fix(ci): make server-capability-routing guard DB-free
The new test:server-capability-routing check failed in CI's Contract
verifiers job: importing getCapabilityServerTypes from
serverResolver.ts transitively pulled in ./prisma, which throws
'DATABASE_URL is missing' at module-load time -- this job intentionally
never sets DATABASE_URL for its DB-free static checks (same reason
textProviderService.ts's own guard avoids a static serverResolver
import).

Split the pure capability->serverType mapping out into its own
server/utils/serverCapabilities.ts with zero runtime Prisma dependency
(type-only Server import). serverResolver.ts now imports from it;
the guard imports it directly, verified locally with DATABASE_URL
unset to reproduce the CI job exactly.
